+++ iceoryx_binding_c/test/moduletests/test_subscriber.cpp
@@ -431,7 +431,7 @@ TEST(iox_sub_options_test, subscriberOptionsAreInitializedCorrectly)
     sut.historyRequest = 73;
     sut.nodeName = "Dr.Gonzo";
     sut.subscribeOnCreate = false;
-    sut.queueFullPolicy = QueueFullPolicy_BLOCK_PUBLISHER;
+    sut.queueFullPolicy = QueueFullPolicy_BLOCK_PRODUCER;
 
     SubscriberOptions options;
     // set subscribeOnCreate to the opposite of the expected default to check if it gets overwritten to default